requirements.txt
pillow
google-generativeai
python-dotenv
retry